+++ ext/io/event/selector/epoll.c
@@ -870,6 +870,8 @@ void * select_internal(void *_arguments) {
 
 static
 void select_internal_without_gvl(struct select_arguments *arguments) {
+	IO_Event_Selector_before_waiting(arguments->selector);
+	
 	arguments->selector->blocked = 1;
 	rb_thread_call_without_gvl(select_internal, (void *)arguments, RUBY_UBF_IO, 0);
 	arguments->selector->blocked = 0;